Music, probably broken though
oh yeah i pushed this change while INTERSTATE. because i am obviosuly such a legendary programmer i will be giving myself a 100% payrise.

so the music is probably a bit broken? i haven't extensively tested it. also it crackles and clips a little bit (might just be my headphones, but didnt we have a similar issue with space game and u fixxed it somehow?) so PLS HELP